requirements.txt
Flask==0.10.1
Jinja2==2.8
nltk==3.0.5
PyYAML==3.11
tweepy==3.4.0
BeautifulSoup==3.2.1
feedparser==5.2.1